Trade in cards as soon as necessary
Until now, we only traded cards in when the engine forced us
to. This meant that we did all of the placing before
receiving the additional troops, and needing to place them
on suboptimal countries.
This commit fixes this by detecting if we need to trade in
cards, and doing so as soon as possible.