Cache RiskGame.getBoard() calls
RiskGame.getBoard() is a very expensive operation. As I found out
through profiling, it is called very often and takes up so much
resources we run into an OOM exception. In this commit, I try to cache
RiskGame.getBoard() as greedy as possible. As a result, a full game now
can be played.
This commit also contains one small fix where we tried to access a
closed stream.